% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSVS.R
\name{FIA2NRCS}
\alias{FIA2NRCS}
\title{Convert species codes from FIA number to NRCS code}
\usage{
FIA2NRCS(Data)
}
\arguments{
\item{MyData}{DataFrame with tree information}
}
\value{
Data frame with converted species codes
}
\description{
Function not implemented yet.
}
\examples{
FIA2NRCS( MyData )
}
\author{
Jim McCarter \email{jim.mccarter@rayonier.com}
}
