% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSVS.R
\name{NRCS2FIA}
\alias{NRCS2FIA}
\title{Convert species codes from NRCS code to FIA number}
\usage{
NRCS2FIA(DataFrame)
}
\arguments{
\item{MyData}{DataFrame with tree information}
}
\description{
Details
}
\examples{
NRCS2FIA( MyData )
}
