% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSVS.R
\docType{package}
\name{rSVS-package}
\alias{rSVS-package}
\title{A package for stand level visualization using the Stand Visualization System (SVS, Robert J. McGaughey, USDA Forest Service, PNW Research Station).}
\description{
This rSVS package provides and interface to perform SVS visualiations from R.
}
\details{
The package includes the following functions:
\itemize{
    \item SVS()             - main function for performing visualziations
    \item SVS_Environment() - check package environment and returns path to components
    \item SVS_Example()     - show reginal example visulizations
    \item SVS_ExampleData() - generate stand data for visulizations
    \item SVS_Species()     - list known species
    \item svsfiles_clean()  - clean out svsfiles folder containing temporary files for visualizations
    \item FIA2NRCS()        - convert species codes from FIA # to NRCS code
    \item NRCS2FIA()        - convert species codes from NRCS code to FIA #
}

The SVS() function supports multple data format types, including:

\itemize{
    \item FMD data frame - list of header, measurement, and treelist slots
    \item LMS Object - list of stand, measurement, and treelist slots
    \item StandObject - list of header, treelist, and management slots
    \item SVScsv - data frame or .csv file containing: Species, PlantID, PlntClass, CrwnClass, TreeStat, DBH, Height, LAng, FAng, EndDia, CRad1, CRat1,
                   CRad2, CRat2, CRad3, CRat3, CRad4, CRat4, TPA, MarkCode, X, Y, Z
    \item StandViz - data frame or .csv file containing: Stand, Year/Age, Species, TreeNo, Live/Dead, Status, Condition, DBH, Height, CrownRatio, CrownRadius, TPA
    \item StandVizExtended - data frame or .csv file containing: Stand, Year/Age, Species, TreeNo, Live/Dead, Status, Condition, DBH, Height, CrownRatio,
                             CrownRadius, TPA, BrokenHt, Offset, Bearing, Lean, RootWad, X, Y
    \item TBL2SVS data frame - data frame or .csv file containing: Species, DBH, Height, CrownRatio, CrownRadius, Status, PlantClass, CrownClass, TPA
}

FMD data frame:

A FMD data frame contains PlotName in column 3, TreeNo, Species, MeasDate, MeasAge in columns 9-12, and Status, Condition, Damage,
Screen, DBH, Height, CrownRatio in columns 14-21.

LMS object:

A LMS object is a list with $stand, $measurement, and $treelist slots.  The $stand slot must contain $STANDNAME, the $measurement must
contain $MEASDATA where year can be extracted from the first 4 character positions.  The $treelist slot must contain tree, species,
tpa, dbh, height, and cr columns in that order.

StandObject:

A StandObject is a list that contains at a minimum $header and #treelist slots.  The $header slot must contain $standid and
$ysp (years since planting) fields.  The $treelist slopt must contain treeid, species, tpa, dbh, height, and cr columns in
that order.

StandViz:

A data frame or .csv file containing: Stand, Year/Age, Species, TreeNo, Live/Dead, Status, Condition, DBH, Height, CrownRatio, CrownRadius, TPA

StandVizExtended:

A data frame or .csv file containing: Stand, Year/Age, Species, TreeNo, Live/Dead, Status, Condition, DBH, Height, CrownRatio, CrownRadius, TPA,
BrokenHt, Offset, Bearing, Lean, RootWad, X, Y.

This format can be used for visualizing mapped stand data and supports a variety of enhancedments that allow for improved visualizations of
forest health conditions, including broken trees (BrokenHt, Offset, Bearing), dead top trees(Condition, BrokenHt), blown down trees (RootWad).
If TPA > 1 then coordinates are generated for the stand.  To provide a mapped stand all TPA values must be 1.0 and X,Y locations filled in.

TBL2SVS:

A data frame or .csv file containing: Species, DBH, Height, CrownRatio, CrownRadius, Status, PlantClass, CrownClass, TPA [, X, Y, MarkCode, FAng, SDia]

Status Codes:
\tabular{lll}{
  \tab 0 or 10 \tab Plant is cut, has branches, on ground \cr
  \tab 1 or 11 \tab plant is standing \cr
  \tab 2 or 12 \tab stump \cr
  \tab 3 or 13 \tab plant is cut, has no branches, on ground \cr
}

This package includes a number of executable programs that will be run as part of the package.
This limits where the package can be hosted (e.g NOT on CRAN).

The package understands a number of data formats and can create visualizaions for single or multiple
stands and years based on what information is contained in the specific data format. Files for
visualzations are created in the svsfiles folder in the current working directory for R. These
temporary files (*.asc, *.bmp, *.csv, *.png, *.SVS, *.opt) are intermediate files used for creating
the visualizations. This folder can be cleaned out using the \strong{clean_svsfiles()} function.

NOTE: SVS is a Windows only program, therefor limiting this package to only work on Windows
computers.
}
