% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSVS.R
\name{SVS_Environment}
\alias{SVS_Environment}
\title{Check SVS environment and return path to components}
\usage{
SVS_Environment(Component = "All", Verbose = FALSE)
}
\arguments{
\item{Component}{which part of the enviroment to check, default all}

\item{Verbose}{echo status messages as environment is being examined}
}
\value{
Messasge messages return and echoed on console
}
\description{
SVS_Enviroment() checks the package enviroment and alternatively will install a python distribution
}
\details{
Details go here
}
\examples{
SVS_Environment( 'Python' )
}
\author{
Jim McCarter \email{jim.mccarter@rayonier.com}
}
