% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSVS.R
\name{SVS_Species}
\alias{SVS_Species}
\title{List known species codes}
\usage{
SVS_Species()
}
\value{
DataFrame with species names and codes
}
\description{
List known species codes.
}
\details{
This function simply reads and returns the list of known species codes distributed with the rSVS
package (rSVS_Species.csv). This file is also used as the basis for species translation by the FIA2NRCS() and
NRCS2FIA() funtions.  The file should be syncronized with the FIA.TFM and NRCS.TFM files, but
this is not guaranteed.

The rSVS package supports FIA number and NRCS alphabetic codes.  The displayed table will include FIA, NRCS, Genus,
Species, Common, and Comment columns.
}
\examples{
SVS_Species()
head(SVS_Species())
length(SVS_Species()$FIA)
}
