% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSVS.R
\name{svsfiles}
\alias{svsfiles}
\title{Function to list and clean out svsfiles folder}
\usage{
svsfiles(clean = FALSE)
}
\arguments{
\item{clean}{}
}
\description{
List of clean out svsfiles folder
}
\details{
Use the svsfiles() function to manage files in the svsfiles temporary folder.
Using the function with no arguments to view a list of files in the folder.  
Use svsfiles(clean=TRUE) to remove files from the folder
}
\examples{
svsfiles()
svsfiles(clean=TRUE)
}
\author{
James McCarter \email{jim.mccarter@rayonier.com}
}
