% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSVS.R
\name{Translate_NRCS2FIA}
\alias{Translate_NRCS2FIA}
\title{Translate species codes from NRCS code to FIA number}
\usage{
Translate_NRCS2FIA(Data)
}
\arguments{
\item{MyData}{DataFrame with tree information}
}
\value{
Data frame with converted species codes
}
\description{
Function not implemented yet.
}
\examples{
Translate_NRCS2FIA( MyData )
}
\author{
Jim McCarter \email{jim.mccarter@rayonier.com}
}
