% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSVS.R
\name{FIA2NRCS}
\alias{FIA2NRCS}
\title{Convert species codes from FIA number to NRCS code}
\usage{
FIA2NRCS(DataFrame)
}
\arguments{
\item{MyData}{DataFrame with tree information}
}
\description{
Details
}
\examples{
FIA2NRCS( MyData )
}
