% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSVS.R
\name{NRCS2FIA}
\alias{NRCS2FIA}
\title{Convert species codes from NRCS code to FIA number}
\usage{
NRCS2FIA(DataFrame)
}
\arguments{
\item{MyData}{DataFrame with tree information}
}
\value{
Data frame with converted species codes
}
\description{
Function not implemented yet.
}
\examples{
NRCS2FIA( MyData )
}
\author{
Jim McCarter \email{jim.mccarter@rayonier.com}
}
