% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSVS.R
\docType{package}
\name{rSVS-package}
\alias{rSVS-package}
\title{A package for stand level visualization using the Stand Visualization System (SVS, Robert J. McGaughey, USDA Forest Service, PNW Research Station).}
\description{
This rSVS package provides an interface to perform SVS visualiations from R.
}
\details{
The package includes the following functions:
\itemize{
    \item svs()                - main function for performing visualziations
    \item svsdata()            - generate stand data for visualizations
    \item svsfiles()           - list or clean out svsfiles folder containing temporary files for visualizations
    \item SVS_Environment()    - check package environment and returns path to components
    \item SVS_Example()        - show reginal example visualizations
    \item SVS_Species()        - list known species
    \item Translate_FIA2NRCS() - convert species codes from FIA # to NRCS code
    \item Translate_NRCS2FIA() - convert species codes from NRCS code to FIA #
}

The package supports either species identified by FIA # or NRCS Plants Database code. The default
behavior can be configured by setting either hidden variable ".UseFIA <- TRUE" or ".UseNRCS <- TRUE".
NOTE: Depending on the source data, setting the incorrect treeform file will cause trees to be
displayed as an Unknown tree species (e.g. red)

The SVS() function supports multiple data format types, including:
\itemize{
    \item FMD data frame - a data frame with at least TreeKey, Species, MeasDate, Status, DBH, Height, CrownRatio, TPA
    \item LMS Object - a list with stand, measurement, and treelist slots
    \item StandObject - a list with header, treelist, and management slots
    \item SVScsv - a data frame or .csv file containing: Species, PlantID, PlntClass, CrwnClass, TreeStat, DBH, Height, LAng, FAng, EndDia, CRad1, CRat1,
                   CRad2, CRat2, CRad3, CRat3, CRad4, CRat4, TPA, MarkCode, X, Y, Z
    \item StandViz - a data frame or .csv file containing: Stand, Year/Age, Species, TreeNo, Live/Dead, Status, Condition, DBH, Height, CrownRatio, CrownRadius, TPA
    \item StandVizExtended - a data frame or .csv file containing: Stand, Year/Age, Species, TreeNo, Live/Dead, Status, Condition, DBH, Height, CrownRatio,
                             CrownRadius, TPA, BrokenHt, Offset, Bearing, Lean, RootWad, X, Y
    \item TBL2SVS data frame - a data frame or .csv file containing: Species, DBH, Height, CrownRatio, CrownRadius, Status, PlantClass, CrownClass, TPA
}

FMD data frame details:

A FMD data frame contains PlotName in column 3, TreeNo, Species, MeasDate, MeasAge in columns 9-12, and Status, Condition, Damage,
Screen, DBH, Height, CrownRatio in columns 14-21.

LMS object details:

A LMS object is a list with $stand, $measurement, and $treelist slots.  The $stand slot must contain $STANDNAME, the $measurement must
contain $MEASDATA where year can be extracted from the first 4 character positions.  The $treelist slot must contain tree, species,
tpa, dbh, height, and cr columns in that order.

StandObject details:

A StandObject is a list that contains at a minimum $header and #treelist slots.  The $header slot must contain $standid and
$ysp (years since planting) fields.  The $treelist slopt must contain treeid, species, tpa, dbh, height, and cr columns in
that order.

StandViz details:

A data frame or .csv file containing: Stand, Year/Age, Species, TreeNo, Live/Dead, Status, Condition, DBH, Height, CrownRatio, CrownRadius, TPA

StandVizExtended details:

A data frame or .csv file containing: Stand, Year/Age, Species, TreeNo, Live/Dead, Status, Condition, DBH, Height, CrownRatio, CrownRadius, TPA,
BrokenHt, Offset, Bearing, Lean, RootWad, X, Y.

This format can be used for visualizing mapped stand data and supports a variety of enhancedments that allow for improved visualizations of
forest health conditions, including broken trees (BrokenHt, Offset, Bearing), dead top trees(Condition, BrokenHt), blown down trees (RootWad).
If TPA > 1 then coordinates are generated for the stand.  To provide a mapped stand all TPA values must be 1.0 and X,Y locations filled in.

TBL2SVS details:

A data frame or .csv file containing: Species, DBH, Height, CrownRatio, CrownRadius, Status, PlantClass, CrownClass, TPA [, X, Y, MarkCode, FAng, SDia]

Status Codes:
\tabular{lll}{
  \tab 0 or 10 \tab Plant is cut, has branches, on ground \cr
  \tab 1 or 11 \tab plant is standing \cr
  \tab 2 or 12 \tab stump \cr
  \tab 3 or 13 \tab plant is cut, has no branches, on ground \cr
}

The following example input files are provided:
\tabular{lcl}{
  - BottomlandHardwood.svs   \tab - \tab SVS format file of mixed bottomland hardwood from southeast \cr
  - Douglas-fir.svs          \tab - \tab SVS format file of Douglas-fir from Pacfic Northwest \cr
  - LodgepolePine.svs        \tab - \tab SVS format file of Lodgepole pine from Rocky Mountains \cr
  - MixedConifer.svs         \tab - \tab SVS format file of Mixed conifer stand from California \cr
  - MontaneOak-Hickory.svs   \tab - \tab SVS format file of Oak/Hickory forest \cr
  - SilverFir-Hemlock.svs    \tab - \tab SVS format file of Pacific silver fir/hemlock stand from Pacfic Northest \cr
  - Redwood.svs              \tab - \tab SVS format file of Redwood stand from Calinornia \cr
  - SouthernPine.svs         \tab - \tab SVS format file of Southern pine stand from southeast \cr
  - SouthernPine.csv         \tab - \tab Same as above, but in SVScsv format \cr
  - SouthernPine2.csv        \tab - \tab Abreviated version of SouthernPine without 4 crown dimensions, but including X,Y coordinates \cr
  - SouthernPine3.csv        \tab - \tab Abreviated version of SouthernPine without X, Y coordinates \cr
  - Spruce-Fir.csv           \tab - \tab SVS format file of Spruce/Fir from Rocky Mountains \cr
}

NOTE: This package includes a number of executable programs that will be run as part of the package,
limiting where the package can be hosted (e.g NOT on CRAN).

The package understands a number of data formats and can create visualizations for single or multiple
stands and years based on what information is contained in the specific data format. Files for
visualizations are created in the svsfiles folder in the current working directory for R. These
temporary files (*.asc, *.bmp, *.csv, *.png, *.SVS, *.opt) are intermediate files used for creating
the visualizations. This folder can be cleaned out using the \strong{svsfiles_clean()} function.

NOTE: SVS is a Windows only program, therefor limiting this package to only work on Windows
computers.
}
