% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSVS.R
\docType{package}
\name{rSVS}
\alias{rSVS}
\title{A packge for stand level visualization using the Stand Visualization System (SVS).}
\description{
This rSVS package provides and interface to perform SVS visualiations from R.
}
\details{
The package includes the following functions:
\itemize{
    \item SVS()             - main function for performing visualziations
    \item SVS_Environment() - show reginal example visulizations
    \item SVS_Example()     - show reginal example visulizations
    \item SVS_Species()     - list known species
    \item FIA2NRCS()        - convert species codes from FIA # to NRCS code
    \item NRCS2FIA()        - convert species codes from NRCS code to FIA #
}

This package includes a number of executable programs that will be run as part of the package.
This limits where the package can be hosted (e.g not on CRAN).

NOTE: SVS is a Windows only program, therefor limiting this package to only work on Windows
computers.
}
