% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSVS.R
\name{SVS}
\alias{SVS}
\title{Visualize stand using the Stand Visualization System (SVS)}
\usage{
SVS(
  data,
  sheet = FALSE,
  output = "svs",
  clumped = FALSE,
  random = TRUE,
  row = FALSE,
  uniform = FALSE,
  randomness = NULL,
  clumpiness = NULL,
  clumpratio = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{compatible data frame or string containing filename with path (see details)}

\item{output}{what and were to product output (SVS | BITMAP | WEB | CSV )}

\item{clumped}{if TRUE generate clumped coordinates}

\item{random}{if TRUE generate random coordiantes}

\item{row}{if TRUE generate coordinates for rows (plantation)}

\item{uniform}{if TRUE generate uniform coordinates}

\item{clumpiness}{adjust clump strength}

\item{clumpratio}{adjust number/size of clumps}

\item{verbose}{turn on verbose output}

\item{randommess}{control "noise" of coordinates generated}
}
\description{
The SVS() function will create stand level visualizations of data frames containing appropriate information.  The
function has the abillity to generate coordinates if they are not provided.
}
\details{
StandViz internal format:
stand, year, species, treeno, x, y, dbh, height, crownratio, crownradius, tpa, live, status, condition, svsstatus, bearing, brokenht, brokenoffset, dmr, leanangle, rootwad

rsvs data frame format:
stand, year, treeno, species, dbh, height, crownratio, crownradius, tpa, x, y, live, status, condition, (svsstatus, brokenht, brokenoffset, bearing, dmr, leanangle, rootwad)

Live/Dead: live or l|dying|dead or d|stump or s
Status: standing or s|broken or b|brokentop|deadtop|down or d
Condition: Live:  1 or dominant or d|2 or codominant or c|3 or intermediate or i|4 or suppressed
           Dying: 1|2|3
           Dead:  1|2|3|4|5
}
\examples{
SVS( d )
SVS( '../MyFiles/Stand1.csv' )
SVS( '../MyFiles/Stand2.xlsx', Sheet='Sheet1' )
SVS( d, random=true )    # visualize stand in svs using random tree locations
SVS( d, row=true )       # visualize stand in svs using rows
}
\author{
James Mccarter \email{jim.mccarter@rayonier.com}
}
