% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSVS.R
\name{SVS_Demo}
\alias{SVS_Demo}
\title{Demonstrate Stand Visualiztion on several stand types}
\usage{
SVS_Demo(Stand = NULL)
}
\arguments{
\item{StandType}{Name of stand to display}
}
\description{
Display one of several stand types using example SVS files includes with the package.
}
\details{
The list of stand types includes:
\itemize{
   \item BottomlandHardwood
   \item Douglas-fir
   \item LodgepolePine
   \item MixedConifer
   \item MontaneOak-Hickory
   \item PacificSilverFir-Hemlock
   \item Redwood
   \item SouthernPine
   \item Spruce-Fir
}
}
\examples{
SVS_Demo( 'SouthernPine' )
SVS_Demo( 'Douglas-fir' )
}
\author{
Jim McCarter \email{jim.mccarter@rayonier.com}
}
