% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSVS.R
\name{SVS_Example}
\alias{SVS_Example}
\title{Demonstrate Stand Visualiztion on several stand types}
\usage{
SVS_Example(Example = NULL)
}
\arguments{
\item{Example}{Name of stand/stand type to display}
}
\description{
Display one of several stand types using example SVS files included with package.
}
\details{
The list of available stand types include:
\itemize{
   \item BottomlandHardwood
   \item Douglas-fir
   \item LodgepolePine
   \item MixedConifer
   \item MontaneOak-Hickory
   \item PacificSilverFir-Hemlock
   \item Redwood
   \item SouthernPine
   \item Spruce-Fir
}
}
\examples{
SVS_Example( 'SouthernPine' )
SVS_Example( 'Douglas-fir' )
SVS_Example()                   # gives list of possible options
}
\author{
Jim McCarter \email{jim.mccarter@rayonier.com}
}
