% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSVS.R
\name{SVS_Species}
\alias{SVS_Species}
\title{List species codes}
\usage{
SVS_Species()
}
\value{
DataFrame with species names and codes
}
\description{
List known species codes.
}
\details{
This function simply reads and returns the list of known species codes distributed with the rSVS
package (rSVS_Species.csv). This file is also used as the basis for species translation by the
Translate_FIA2NRCS() and Translate_NRCS2FIA() funtions.

The rSVS package supports FIA number and NRCS alphabetic codes. The displayed table will include
FIA, NRCS, Genus, Species, Common, and Comment, FIA.trf, NRCS.trf, FVS.trf, FVSSpCode, FVS.East,
and FVS.West.

The FIA.trf and NRCS.trf columns should list the species that currently exist in each SVS tree
form file (NOTE: this species list should be synchronized with FIA.trf and NRCS.trf, but this is
not guaranteed).

The FVS.trf, FVSSpCode, FVS.East, and FVS.West columns are used for species audits and keeping
track of the source of tree form definitions (largely from FIA variant tree form files from the
original SVS distribution).  Note that the tree form definitions have been exanded to handle more
live and dead classes for improved visualizations of forest health conditions.
}
\examples{
Species <- SVS_Species()            # store species list to Species
names(SVS_Species)                  # column names in returned species list
head(SVS_Species())                 # first 6 species records
length(SVS_Species()$FIA)           # number of species records
length(unique(SVS_Species()$FIA))   # number of FIA # known
sort(unique(SVS_Species()$FIA))     # list of FIA #'s known
length(unique(SVS_Species()$NRCS))  # number of NRCS codes known
}
